requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
cloudinary==1.28.0
dj-database-url==0.5.0
Django==3.2.9
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
